GCP: Introduce get_cell_properties_nonperiodic()
This works like get_cell_properties() but without assuming a periodic lattice. Instead init_YTZP and init_twiss are given as argument. This required some changes, so the orbit_data returned is now a subclass of numpy.ndarray (hopefully this will not cause issues). Some of the columns returned are different. Some of the functions get_cell_tracks(), plot_cell_tracks() and get_twiss_params() have been updated to handle this.